# Fix critical 21x performance regression in ripgrep integration