--- a/src/search/ripgrep_searcher.rs
+++ b/src/search/ripgrep_searcher.rs
@@ -1,97 +1,30 @@
 use anyhow::{Context, Result};
-use grep_regex::{RegexMatcher, RegexMatcherBuilder};
-use grep_searcher::{Searcher, SearcherBuilder, Sink, SinkMatch};
 use regex::RegexSet;
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
 use std::time::Instant;
 
-/// Configuration for I/O optimization strategies
-#[derive(Debug, Clone)]
-pub struct IOConfig {
-    pub buffer_size: usize,
-    pub use_memory_map: bool,
-    pub max_file_size: u64,
-    pub streaming_threshold: u64,
-    pub large_file_threshold: u64,
-    pub force_mmap_threshold: u64,
-}
-
-impl Default for IOConfig {
-    fn default() -> Self {
-        IOConfig {
-            buffer_size: 64 * 1024,                 // 64KB default buffer
-            use_memory_map: true,                   // Enable memory mapping by default
-            max_file_size: 100 * 1024 * 1024,       // 100MB max file size
-            streaming_threshold: 10 * 1024 * 1024,  // 10MB streaming threshold
-            large_file_threshold: 50 * 1024 * 1024, // 50MB large file threshold
-            force_mmap_threshold: 20 * 1024 * 1024, // 20MB force memory mapping
-        }
-    }
-}
-
-impl IOConfig {
-    /// Create a configuration optimized for large files
-    pub fn large_files() -> Self {
-        IOConfig {
-            buffer_size: 256 * 1024, // 256KB buffer for large files
-            use_memory_map: true,
-            max_file_size: 1024 * 1024 * 1024, // 1GB max file size
-            streaming_threshold: 5 * 1024 * 1024, // 5MB streaming threshold
-            large_file_threshold: 100 * 1024 * 1024, // 100MB large file threshold
-            force_mmap_threshold: 50 * 1024 * 1024, // 50MB force memory mapping
-        }
-    }
-
-    /// Create a configuration optimized for small files
-    pub fn small_files() -> Self {
-        IOConfig {
-            buffer_size: 8 * 1024,                 // 8KB buffer for small files
-            use_memory_map: false,                 // Disable memory mapping for small files
-            max_file_size: 10 * 1024 * 1024,       // 10MB max file size
-            streaming_threshold: 1024 * 1024,      // 1MB streaming threshold
-            large_file_threshold: 5 * 1024 * 1024, // 5MB large file threshold
-            force_mmap_threshold: u64::MAX,        // Never force memory mapping
-        }
-    }
-}
-
-/// High-performance ripgrep-based searcher that replaces the custom RegexSet implementation
-/// This provides significant performance improvements through:
-/// - Native SIMD optimizations via the Teddy algorithm
-/// - Optimized memory mapping and streaming I/O
+/// High-performance RegexSet-based searcher for fast file pattern matching
+/// This provides optimal performance through:
+/// - Pre-compiled RegexSet for multi-pattern efficiency  
+/// - Simple file I/O without unnecessary abstraction layers
 /// - Parallel processing capabilities
-/// - Advanced encoding detection and compression support
+/// - Thread-safe design for concurrent access
 pub struct RipgrepSearcher {
-    patterns: Vec<String>,
-    enable_simd: bool,
     debug_mode: bool,
-    io_config: IOConfig,
     regex_set: RegexSet,
 }
 
 impl RipgrepSearcher {
     /// Create a new RipgrepSearcher with optimized settings
-    pub fn new(patterns: &[String], enable_simd: bool) -> Result<Self> {
-        Self::with_config(patterns, enable_simd, IOConfig::default())
-    }
-
-    /// Create a new RipgrepSearcher with custom I/O configuration
-    pub fn with_config(
-        patterns: &[String],
-        enable_simd: bool,
-        io_config: IOConfig,
-    ) -> Result<Self> {
+    pub fn new(patterns: &[String], _enable_simd: bool) -> Result<Self> {
         let debug_mode = std::env::var("DEBUG").unwrap_or_default() == "1";
 
         if debug_mode {
             println!(
-                "DEBUG: Creating RipgrepSearcher with {} patterns",
+                "DEBUG: Creating fast RegexSet searcher with {} patterns",
                 patterns.len()
             );
-            println!("DEBUG: SIMD enabled: {enable_simd}");
-            println!("DEBUG: I/O config: {io_config:?}");
         }
 
         // Pre-compile RegexSet for efficient pattern matching
@@ -101,71 +34,13 @@ impl RipgrepSearcher {
             .context("Failed to build RegexSet during initialization")?;
 
         Ok(RipgrepSearcher {
-            patterns: patterns.to_vec(),
-            enable_simd,
             debug_mode,
-            io_config,
             regex_set,
         })
     }
 
-    /// Create a matcher and searcher for thread-local use
-    fn create_matcher_searcher(&self) -> Result<(RegexMatcher, Searcher)> {
-        // Combine patterns into a single alternation for initial filtering
-        let combined_pattern = if self.patterns.len() == 1 {
-            self.patterns[0].clone()
-        } else if self.patterns.is_empty() {
-            ".*".to_string() // Default fallback pattern
-        } else {
-            format!("(?:{})", self.patterns.join("|"))
-        };
-
-        if self.debug_mode {
-            println!("DEBUG: Combined pattern: {combined_pattern}");
-        }
-
-        // Build regex matcher with optimizations
-        let mut matcher_builder = RegexMatcherBuilder::new();
-        matcher_builder
-            .case_insensitive(true)
-            .multi_line(true)
-            .unicode(true);
-
-        // Enable advanced SIMD optimizations if requested
-        if self.enable_simd {
-            // The Teddy algorithm provides 2-10x speedup for multi-pattern matching
-            matcher_builder.dfa_size_limit(10 * (1 << 20)); // 10MB DFA size limit
-        }
-
-        let matcher = matcher_builder
-            .build(&combined_pattern)
-            .context("Failed to build regex matcher")?;
-
-        // Build searcher with memory mapping and streaming optimizations
-        let mut searcher_builder = SearcherBuilder::new();
-        searcher_builder
-            .line_number(true)
-            .binary_detection(grep_searcher::BinaryDetection::quit(b'\x00'))
-            .multi_line(false); // Process line by line for better memory efficiency
-
-        // Configure memory mapping based on I/O settings and file size heuristics
-        if self.io_config.use_memory_map {
-            // Use automatic memory mapping decision for optimal performance
-            searcher_builder.memory_map(unsafe { grep_searcher::MmapChoice::auto() });
-        } else {
-            searcher_builder.memory_map(grep_searcher::MmapChoice::never());
-        }
-
-        // Note: buffer_size configuration is handled internally by ripgrep
-        // We can't directly set buffer size through the public API
-        // The buffer size is automatically optimized based on file size and memory mapping
-
-        let searcher = searcher_builder.build();
-
-        Ok((matcher, searcher))
-    }
-
     /// Search a single file and return term matches with line numbers
+    /// This uses a fast RegexSet-based approach for maximum performance
     pub fn search_file(
         &self,
         file_path: &Path,
@@ -175,73 +50,84 @@ impl RipgrepSearcher {
         let mut term_map = HashMap::new();
 
         if self.debug_mode {
-            println!("DEBUG: Searching file with ripgrep: {file_path:?}");
+            println!("DEBUG: Searching file with optimized RegexSet: {file_path:?}");
         }
 
-        // Check file size before processing
-        let metadata = std::fs::metadata(file_path).context("Failed to get file metadata")?;
+        // Define a reasonable maximum file size (same as old implementation)
+        const MAX_FILE_SIZE: u64 = 1024 * 1024; // 1MB
 
-        if metadata.len() > self.io_config.max_file_size {
+        // Check file metadata and resolve symlinks before reading
+        let resolved_path = match std::fs::canonicalize(file_path) {
+            Ok(path) => path,
+            Err(e) => {
+                if self.debug_mode {
+                    println!("DEBUG: Error resolving path for {file_path:?}: {e:?}");
+                }
+                return Err(anyhow::anyhow!("Failed to resolve file path: {}", e));
+            }
+        };
+
+        // Get file metadata to check size and file type
+        let metadata = match std::fs::metadata(&resolved_path) {
+            Ok(meta) => meta,
+            Err(e) => {
+                if self.debug_mode {
+                    println!("DEBUG: Error getting metadata for {resolved_path:?}: {e:?}");
+                }
+                return Err(anyhow::anyhow!("Failed to get file metadata: {}", e));
+            }
+        };
+
+        // Check if the file is too large
+        if metadata.len() > MAX_FILE_SIZE {
             if self.debug_mode {
                 println!(
-                    "DEBUG: Skipping large file: {:?} ({} bytes > {} bytes limit)",
-                    file_path,
+                    "DEBUG: Skipping file {:?} - file too large ({} bytes > {} bytes limit)",
+                    resolved_path,
                     metadata.len(),
-                    self.io_config.max_file_size
+                    MAX_FILE_SIZE
                 );
             }
-            return Ok(term_map);
+            return Err(anyhow::anyhow!(
+                "File too large: {} bytes (limit: {} bytes)",
+                metadata.len(),
+                MAX_FILE_SIZE
+            ));
         }
 
-        // Log I/O strategy decisions based on file size
-        if self.debug_mode {
-            if metadata.len() > self.io_config.large_file_threshold {
-                println!(
-                    "DEBUG: Large file detected: {:?} ({} bytes > {} bytes threshold) - using optimized large file strategy",
-                    file_path,
-                    metadata.len(),
-                    self.io_config.large_file_threshold
-                );
-            } else if metadata.len() > self.io_config.force_mmap_threshold {
-                println!(
-                    "DEBUG: Medium file detected: {:?} ({} bytes > {} bytes threshold) - forcing memory mapping",
-                    file_path,
-                    metadata.len(),
-                    self.io_config.force_mmap_threshold
-                );
-            } else if metadata.len() > self.io_config.streaming_threshold {
-                println!(
-                    "DEBUG: Using streaming I/O for file: {:?} ({} bytes > {} bytes threshold)",
-                    file_path,
-                    metadata.len(),
-                    self.io_config.streaming_threshold
-                );
+        // Read the file content with proper error handling (simple and fast)
+        let content = match std::fs::read_to_string(&resolved_path) {
+            Ok(content) => content,
+            Err(e) => {
+                if self.debug_mode {
+                    println!(
+                        "DEBUG: Error reading file {:?}: {:?} (size: {} bytes)",
+                        resolved_path,
+                        e,
+                        metadata.len()
+                    );
+                }
+                return Err(anyhow::anyhow!("Failed to read file: {}", e));
             }
-        }
+        };
 
-        // Create matcher and searcher for this thread
-        let (matcher, mut searcher) = self.create_matcher_searcher()?;
+        // Process each line (fast in-memory processing)
+        for (line_number, line) in content.lines().enumerate() {
+            // Skip lines that are too long
+            if line.len() > 2000 {
+                if self.debug_mode {
+                    println!(
+                        "DEBUG: Skipping line {} in file {:?} - line too long ({} characters)",
+                        line_number + 1,
+                        file_path,
+                        line.len()
+                    );
+                }
+                continue;
+            }
 
-        // Create a sink to collect matches
-        let matches = Arc::new(Mutex::new(Vec::new()));
-        let matches_clone = Arc::clone(&matches);
-
-        let sink = RipgrepSink::new(matches_clone);
-
-        // Perform the search using ripgrep's optimized engine
-        searcher
-            .search_path(&matcher, file_path, sink)
-            .with_context(|| format!("Failed to search file: {file_path:?}"))?;
-
-        // Process matches to build term map with deterministic ordering
-        let mut collected_matches = matches.lock().unwrap().clone();
-        // Sort matches by line number for deterministic processing
-        collected_matches.sort_by_key(|m| m.line_number);
-
-        // Use pre-compiled RegexSet for efficient multi-pattern matching
-        for line_match in collected_matches.iter() {
-            // Use RegexSet to efficiently find which patterns match this line
-            let matches = self.regex_set.matches(&line_match.line_content);
+            // Use pre-compiled RegexSet for efficient multi-pattern matching
+            let matches = self.regex_set.matches(line);
             if matches.matched_any() {
                 // For each matched pattern, map to corresponding term indices
                 for pattern_idx in matches.iter() {
@@ -250,7 +136,7 @@ impl RipgrepSearcher {
                             term_map
                                 .entry(term_idx)
                                 .or_insert_with(HashSet::new)
-                                .insert(line_match.line_number);
+                                .insert(line_number + 1); // Convert to 1-based line numbers
                         }
                     }
                 }
@@ -260,41 +146,40 @@ impl RipgrepSearcher {
         if self.debug_mode {
             let duration = start_time.elapsed();
             println!(
-                "DEBUG: Ripgrep search completed for {:?} in {:?} - found {} matches",
+                "DEBUG: Fast RegexSet search completed for {:?} in {:?} - found {} unique terms",
                 file_path,
                 duration,
-                collected_matches.len()
+                term_map.len()
             );
         }
 
         Ok(term_map)
     }
 
-    /// Search multiple files in parallel using ripgrep with deterministic ordering
+    /// Search multiple files in parallel using fast RegexSet-based approach
     pub fn search_files_parallel(
         &self,
         file_paths: &[PathBuf],
         pattern_to_terms: &[HashSet<usize>],
     ) -> Result<HashMap<PathBuf, HashMap<usize, HashSet<usize>>>> {
         use rayon::prelude::*;
-        use std::collections::BTreeMap;
 
         let start_time = Instant::now();
 
         if self.debug_mode {
             println!(
-                "DEBUG: Starting parallel ripgrep search on {} files",
+                "DEBUG: Starting parallel RegexSet search on {} files",
                 file_paths.len()
             );
         }
 
-        // Use par_iter().map() to collect results in deterministic order
-        // This preserves the original file order from file_paths
+        // Use par_iter().filter_map() for parallel processing
+        // The searcher instance is thread-safe, so we can reuse it
         let results: Vec<(PathBuf, HashMap<usize, HashSet<usize>>)> = file_paths
             .par_iter()
             .filter_map(|file_path| {
-                // Reuse the existing searcher instance (thread-safe)
-                // The search_file method creates thread-local Matcher and Searcher instances
+                // Reuse the shared searcher instance - it's thread-safe
+                // The search_file method uses only simple file I/O and RegexSet matching
                 match self.search_file(file_path, pattern_to_terms) {
                     Ok(term_map) => {
                         if !term_map.is_empty() {
@@ -313,63 +198,20 @@ impl RipgrepSearcher {
             })
             .collect();
 
-        // Convert to HashMap while preserving deterministic order
-        // Use BTreeMap for consistent iteration order, then convert to HashMap
-        let mut ordered_results = BTreeMap::new();
-        for (file_path, term_map) in results {
-            ordered_results.insert(file_path, term_map);
-        }
-
+        // Convert to HashMap (order doesn't matter for final results)
         let final_results: HashMap<PathBuf, HashMap<usize, HashSet<usize>>> =
-            ordered_results.into_iter().collect();
+            results.into_iter().collect();
 
         if self.debug_mode {
             let duration = start_time.elapsed();
             println!(
-                "DEBUG: Parallel ripgrep search completed in {:?} - found matches in {} files",
+                "DEBUG: Parallel RegexSet search completed in {:?} - found matches in {} files",
                 duration,
                 final_results.len()
             );
         }
 
         Ok(final_results)
-    }
-}
-
-/// Custom sink implementation for collecting ripgrep matches
-struct RipgrepSink {
-    matches: Arc<Mutex<Vec<LineMatch>>>,
-}
-
-#[derive(Debug, Clone)]
-struct LineMatch {
-    line_number: usize,
-    #[allow(dead_code)]
-    line_content: String,
-}
-
-impl RipgrepSink {
-    fn new(matches: Arc<Mutex<Vec<LineMatch>>>) -> Self {
-        RipgrepSink { matches }
-    }
-}
-
-impl Sink for RipgrepSink {
-    type Error = std::io::Error;
-
-    fn matched(&mut self, _searcher: &Searcher, mat: &SinkMatch<'_>) -> Result<bool, Self::Error> {
-        let line_number = mat.line_number().unwrap_or(0) as usize;
-        let line_content = String::from_utf8_lossy(mat.bytes()).to_string();
-
-        let line_match = LineMatch {
-            line_number,
-            line_content,
-        };
-
-        let mut matches_guard = self.matches.lock().unwrap();
-        matches_guard.push(line_match);
-
-        Ok(true) // Continue searching
     }
 }