## Problem Analysis

The ripgrep integration introduced a massive **21x performance regression** (371ms → 8.10s) that made file searching painfully slow instead of blazingly fast as intended.

## Root Cause

The issue was **overengineering** - using ripgrep's full search pipeline (designed for CLI usage) instead of just leveraging its regex engine:

- **Complex I/O overhead**: Created `Matcher`/`Searcher` instances for each file  
- **Unnecessary abstraction**: Used ripgrep's `Sink` trait and `search_path` API
- **Double regex work**: Both ripgrep's internal regex AND separate `RegexSet`
- **Memory mapping overhead**: Complex I/O configuration for simple file reads
- **Per-file instance creation**: Thread-local searcher creation instead of reusing shared instance

## Solution

Reverted to the **fast RegexSet-based approach** while keeping the pre-compilation benefit:

- ✅ **Simple file I/O**: Direct `std::fs::read_to_string()` - no unnecessary abstraction
- ✅ **Pre-compiled RegexSet**: Efficient multi-pattern matching (kept from ripgrep integration)
- ✅ **In-memory processing**: Fast line-by-line iteration without I/O overhead
- ✅ **Thread-safe reuse**: Shared searcher instance across parallel processing

## Performance Results

- **File searching**: 8.10s → **70ms** (115x speedup\!)
- **vs Original**: 70ms vs 371ms (5x faster than pre-ripgrep baseline)  
- **Total improvement**: Faster than original while keeping regex pre-compilation benefits

## Testing

- ✅ All unit tests pass (96/96)
- ✅ All integration tests pass (7/7, 2 ignored)
- ✅ Performance validated with timing measurements
- ✅ No functional regressions

The fix maintains all existing functionality while delivering the performance that was originally intended from the ripgrep integration.

🤖 Generated with [Claude Code](https://claude.ai/code)